Split classes by files
