+++ SqliteLogger/NullDisposable.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace SqliteLogger
+{
+    internal class NullDisposable : IDisposable
+    {
+        public static readonly IDisposable Instance = new NullDisposable();
+
+        private NullDisposable()
+        {
+
+        }
+
+        public void Dispose()
+        {
+            return;
+        }
+    }
+}
